[RSDK-8124] Extract serve_web_inner to common
